requirements.txt
arrow==1.2.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
blessed==1.19.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coinbase-commerce==1.0.1
cryptography==38.0.1
Django==4.1.2
django-appconf==1.0.5
django-bitfield==2.2.0
django-ckeditor==6.5.1
django-compressor==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-picklefield==3.1
django-q==1.3.9
django-user-agents==0.4.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==2.10
mysql-connector-python==8.0.30
nanoid==2.0.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
paypalrestsdk==1.13.1
protobuf==3.20.1
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.4
rcssmin==1.1.0
redis==3.5.3
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.3
StringGenerator==0.4.4
stripe==4.2.0
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
wcwidth==0.2.5